Use kotest for unit tests. Also move DataSourceTest to a kotlin source directory instead of java.
