+++ gradle/libs.versions.toml
@@ -3,6 +3,7 @@ android = "7.3.1"
 compose = "1.3.2"
 protobuf = "3.21.12"
 kotlin = "1.7.20"
+kotest = "5.5.4"
 
 [libraries]
 androidx-core = { module = "androidx.core:core-ktx", version = "1.9.0" }
@@ -25,6 +26,9 @@ androidx-espresso-core = { module = "androidx.test.espresso:espresso-core", vers
 compose-ui-test-junit4 = { module = "androidx.compose.ui:ui-test-junit4", version.ref = "compose" }
 compose-ui-test-manifest = { module = "androidx.compose.ui:ui-test-manifest", version.ref = "compose" }
 androidx-customview-poolingcontainer = { module = "androidx.customview:customview-poolingcontainer", version = "1.0.0" }
+
+kotest-runner = { module = "io.kotest:kotest-runner-junit5", version.ref = "kotest" }
+kotest-assertions = { module = "io.kotest:kotest-assertions-core", version.ref = "kotest" }
 
 [bundles]
